OrderService 트랜잭션 분리

+++ src/main/java/com/homesweet/homesweetback/domain/order/dto/internal/PendingOrder.java
@@ -14,7 +14,7 @@ public record PendingOrder(
 ) {
     public record PendingOrderItem(
             Long skuId,
-            int quantity,
+            long quantity,
             long price
     ) {}
 }
